Remove 10.1 for now
It's a beta, and it's broken as-is.  Besides that, it's confusing users to have it listed as supported because it's in beta status, and every page upstream has that mentions it includes the disclaimer of "Do not use beta releases on production systems!" (and we don't have a good way to include a similar disclaimer in our documentation).