Add "pinning" to ensure we don't run into issues with mixed dependencies
Specifically, one of the dependencies of mariadb-server has a newer version in `wheezy`, but the dependency in the package was `= 5.5.42+maria-1~wheezy`, meaning it just refused to install.  By pinning the "Origin: MariaDB" repo to be the highest priority, this dependency chain resolves itself properly.

```
libmariadbclient18 : Depends: libmysqlclient18 (= 5.5.42+maria-1~wheezy) but 5.5.43-0+deb7u1 is to be installed
```