Add env variable to skip database initialization
This add a new env variable (MYSQL_SKIP_INIT=1) which skip the
database initialization.

This is useful, when the server act as a Galera node, initializing
the datadir make no sense in that case, as a copy of the datadir
is transfered from another node in the cluster.

See: http://galeracluster.com/documentation-webpages/statetransfer.html