Speed up 10.4+ timezone initialization
MariaDB-10.4 defaulted to Aria for system tables.

This introduced crash safety under the name of "transactional"
that was not previously in MyISAM.

The Aria implementation of checkpointing incurs significant
penality on fuse-overlayfs that occurs significantly in
container environments, especially those without a
/var/lib/mysql volume.

We work around this penality by disabling the crash
safety of timezone tables for the period of timezone
initialization.

Analysis and timings are in https://jira.mariadb.org/browse/MDEV-23326
and local tests show that 10.4 is only 0.8 seconds slower
than 10.3 on startup (6.8 seconds total).

Version specific comments are used to ensure that ALTER TABLE
statements aren't run on < 10.4 server versions.

closes #262