MARIADB_RANDOM_ROOT_PASSWORD - much more random
With https://mariadb.com/kb/en/password-validation/#examples
it may be the case that the random root password fails
to meet the complexity requirements.

We take advantage that both bionic (for 10.2) and focal
onwards include pwgen-2.08 that included the added options
to be secure on multiple ways.

Though not necessary, take away a few characters in
realization that the scripting ability of our users
may not contain the basic escaping needed to process
this added security.