Simplify ARG REPOSITORY / archive.mariadb.org as URL
REPOSITORY can be ARG with value thanks to @yosifkit.

Per https://jira.mariadb.org/browse/MDEV-26122 it
got a bit disturbing to users when we purged off the
OSUOSL mirror. As these files are predominately just
used in the CI and building by our maintainers
hopefully any downside to the build process it gained
in longer term maintainability.