MDEV-27732: Add env MARIADB_MYSQL_LOCALHOST_{USER,GRANTS}
We create a mysql@localhost user under MARIADB_MYSQL_LOCALHOST_USER=1
with its default USAGE privileges.

This gives access to SHOW GLOBAL STATUS/VARIABLES but no other
real access including database visibility or process visibility.

Being a @localhost this restricts access to via the unix socket.

The level of access can be increased controlled by the environment variable
MARIADB_MYSQL_LOCALHOST_GRANTS.

If you are using monitoring replication or processes addition
privileges are required, and the setting of the environment variable
is the comma separated list of grants possible. For the moment,
these are a set of global grants only.

If you share the unix socket location, /var/run/mysqld by default,
as a volume with another container, you have effectively given that
container the mysql@localhost provided they have the same uid map
or can create arbitrary users.

This can be good for things like backup. This would require the
datadir volume as well and the set of privileges
(https://mariadb.com/kb/en/mariabackup-overview/#authentication-and-privileges).

Any grants of UPDATE on mysql database will mean that the
mysql@localhost user can manipulate any other user, potentially
transparently adding unix_socket auth (in 10.4+), and the being able
to gain their privileges.

Grants of CREATE USER, or INSERT on the mysql database allow the
creation of user and privilege escalation.

For these reasons, ALL for MARIADB_MYSQL_LOCALHOST_GRANTS gains a
warning.

Many thanks to Daniel Rudolf for all the reviews and the support
to develop this feature.