skip innodb buffer pool loads/saves on temporary startup/shutdown
This saves a small bit of IO, more so with mariadb-upgrade, and
none of the IO/memory acheives much gain in theses temporary start
situtations.

Also by skipping on the temporary starts, the innodb_buffer_pool_loaded
option in the healthcheck script only gains the "loaded" status
on the final startup in the entrypoint.