MARIADB_MYSQL_LOCALHOST_GRANTS not to imply MARIADB_MYSQL_LOCALHOST_USER
Daniel Rudolf from
https://github.com/MariaDB/mariadb-docker/pull/409#pullrequestreview-873937215:

IMHO a user implies a grant, not necessarily the other way round. Just think of
it as if we made the username (mysql) configurable: You can't grant privileges
without an user, you need the user first. The only reason we can do this is
because the username is hard coded, not because a grant implies its user.

Additionally I'm a bit concerned about unexpected behaviour: Just think about an
user passing --env MARIADB_MYSQL_LOCALHOST_USER= --env MARIADB_MYSQL_LOCALHOST_GRANTS=USAGE.
The user explicitly tells us not to create an user, but the dangling grant
overrules this statement. Sure, no human would ever pass this purposely, but users
don't necessarily execute commands manually, sometimes the options to pass are
computed and/or read from arbitrary sources - which then must all be aware of this
tight coupling.