simplify shutdown in docker_mariadb_upgrade
Initial testing showed that kill -0 sometimes
didn't kill mariadbd enough and that the
startup sill had the aria_control locked.

To work around this additional testing on
aquiring aria_control exclusive locks was
attempted. This also was insufficient and
an additional 2 seconds where added.

Falling back to the capabilites of bash to
use a wait builtin to wait for an examined
process to terminate was a better option
found upon further examination.

For note:

The docker_temp_server_stop uses the
mysqladmin shutdown mechanism. This waits
until the pid file created by the server
is removed. The server code shows this occurs
after the storage engines are shut down.