Allow healthcheck@{127.0.0.1,::1,localhost} to exist to facilitate healthcheck --connect
healthcheck@{127.0.0.1,::1,localhost} users are granted USAGE by default, which
is enough for the non-replication healthchecks in healtcheck.sh.

The env variable MARIADB_HEALTHCHECK_GRANTS can replace USAGE with any
comma separated set of grants.

On initialization a generated password is created and saved in
$DATADIR/.my-healthcheck.cnf along with the server port and socket. If the
command args or default configuration file changes this may become out
of date. Because the password is generated in configuration file the
'#', comment, and '=' characters cannot be part of this password.

The healthcheck.cnf configuration file also sets protocol=tcp to
enforce indirectly that --connect being a standard part of the test. This is
required as starts of the service under --skip-networking should
never be considered healthy.

The healthcheck script also has the --defaults-extra-file set to this
.my-healthcheck.cnf file, if it exists (backwards compatible on
previously created datadirs), so that all new healthcheck invokations
use the authentication here by default.

The compatibility with old instances, without the .my-healthcheck.cnf is
preserved by non setting --defaults-extra-file.

The healthcheck --connect will increment the server status variable Aborted_connects
for each check, however now connection_error* counts are changed.

This also prevents any invalid password errors showing up in the
container log.

Closes #430