healthcheck config on socket/port - omit
The socket/port that where used in the initialization of the datadir
in the /var/lib/mysql may not be the same used at runtime.

This was highlighted in #630 where the ubi container puts a socket
in a different location from the non-ubi container.

On the assumption that any change of default port/socket is going
to be in a server configuration file, preferably with a [client-server]
section, lets omit copying this unnecessary.

This does however mean that command line specified changes to
socket (and port less so) will not be persistent. This means that
podman run mariadb:lts --socket=/tmp/s.sock will not be discoverable
for the healtcheck.sh script.

Closes: #630