EPEL 10 missing jemalloc + pwgen
EPEL is missing jemalloc from 10 (but is in 10.1).
For a few small released we can do without this.

pwgen is also missing from EPEL (bug report
https://bugzilla.redhat.com/show_bug.cgi?id=2370234).
In the mean time we use a perl command as an alternative.

The .test/run.sh is updated to not depend on jemalloc existing.
The jemalloc.1 soname that was in focal is removed from
the test case.